fix: fix -Wextra warnings in tt_metal/fabric/ (r21 batch)
- fabric_builder_context.cpp: cast ChipId to size_t in comparisons
- fabric_tensix_builder_impl.cpp: add missing designated field initializers
- control_plane.cpp: fix sign-compare warnings throughout
- mesh_graph_descriptor.cpp: cast dims_size()/src/dst to uint32_t
- physical_system_discovery.cpp: cast rank to int vs distributed_context->size()
- topology_mapper.cpp: cast peer/source_rank for sign-compare fixes
- topology_solver.cpp: cast neighbor_chip_id to uint32_t vs chip_id
- ccl/ccl_common.cpp: add missing .defines = {} to EthernetConfig init

+++ tt_metal/fabric/physical_system_discovery.cpp
@@ -124,7 +124,7 @@ bool resolve_hostname_uniqueness(
     if (my_rank == controller_rank) {
         std::vector<std::string> hostnames = {};
         hostnames.push_back(get_host_name());
-        for (std::size_t rank = 0; rank < *(distributed_context->size()); rank++) {
+        for (std::size_t rank = 0; static_cast<int>(rank) < *(distributed_context->size()); rank++) {
             if (rank != controller_rank) {
                 std::size_t peer_hostname_size = 0;
                 distributed_context->recv(
@@ -144,7 +144,7 @@ bool resolve_hostname_uniqueness(
         }
         all_hostnames_unique = std::set<std::string>(hostnames.begin(), hostnames.end()).size() == hostnames.size();
 
-        for (std::size_t rank = 0; rank < *(distributed_context->size()); rank++) {
+        for (std::size_t rank = 0; static_cast<int>(rank) < *(distributed_context->size()); rank++) {
             if (rank != controller_rank) {
                 distributed_context->send(
                     tt::stl::Span<std::byte>(
@@ -454,7 +454,7 @@ void exchange_metadata(
 
     if (issue_gather) {
         receiver_ranks.insert(controller_rank);
-        for (std::size_t rank = 0; rank < *(distributed_context->size()); rank++) {
+        for (std::size_t rank = 0; static_cast<int>(rank) < *(distributed_context->size()); rank++) {
             if (rank != controller_rank) {
                 sender_ranks.insert(rank);
             }
